Update Angular example for v20.033
