+++ angular/src/index.html
@@ -2,12 +2,10 @@
 <html lang="en">
 <head>
   <meta charset="utf-8">
-  <title>TradingView Charting Library and Angular 5 Integration Example</title>
+  <title>Angular</title>
   <base href="/">
-
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <script src="https://code.jquery.com/jquery-1.11.2.min.js"></script>
-  <script src="/assets/datafeeds/udf/dist/polyfills.js"></script>
+  <link rel="icon" type="image/x-icon" href="favicon.ico">
   <script src="/assets/datafeeds/udf/dist/bundle.js"></script>
 </head>
 <body>
