--- angular5/e2e/app.po.ts
@@ -1,12 +1,0 @@
-import { browser, by, element } from 'protractor';
-
-export class AppPage {
-  navigateTo() {
-    browser.waitForAngularEnabled(false);
-    return browser.get('/');
-  }
-
-  getHeadingText() {
-    return element(by.css('h1')).getText();
-  }
-}
