--- angular5/src/tsconfig.app.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../out-tsc/app",
-    "baseUrl": "./",
-    "module": "es2015",
-    "types": []
-  },
-  "exclude": [
-    "test.ts",
-    "**/*.spec.ts",
-    "assets/datafeeds/udf/src"
-  ]
-}
